requirements.txt
# EXL FNOL Transcript Analyzer - Requirements
# Industrial-grade dependencies for production deployment

# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # Excel file support

# API & HTTP
requests>=2.31.0
google-genai

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.0

# Type checking (development)
typing-extensions>=4.8.0

# Testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (development)
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Security
python-dateutil>=2.8.0
